# pulsar-rest/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Browser4 WebDriver-Compatible API
  description: |
    A WebDriver-compatible HTTP API with selector-first extensions for browser automation.
    This API provides session management, navigation, element interaction, and event handling
    capabilities.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8182
    description: Local development server
...
paths:
  /session:
    post:
      tags:
        - session
      summary: Create a new browser session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewSessionRequest'
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewSessionResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}:
    get:
      tags:
        - session
      summary: Get session details
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetails'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - session
      summary: Delete a session
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/url:
    post:
      tags:
        - navigation
      summary: Navigate to a URL
      operationId: navigateTo
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUrlRequest'
      responses:
        '200':
          description: Navigation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - navigation
      summary: Get current URL
      operationId: getCurrentUrl
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Current URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/documentUri:
    get:
      tags:
        - navigation
      summary: Get document URI
      operationId: getDocumentUri
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Document URI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/baseUri:
    get:
      tags:
        - navigation
      summary: Get base URI
      operationId: getBaseUri
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Base URI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/{sessionId}/execute/sync:
    post:
      tags:
        - script
      summary: Execute synchronous script
      operationId: executeSync
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptRequest'
      responses:
        '200':
          description: Script result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/execute/async:
    post:
      tags:
        - script
      summary: Execute asynchronous script
      operationId: executeAsync
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptRequest'
      responses:
        '200':
          description: Script result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/{sessionId}/normalize:
    post:
      tags:
        - pulsar
      summary: Normalize a URL
      description: |
        Normalizes a URL with optional load arguments. Returns the normalized URL
        and parsed arguments.
      operationId: normalize
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormalizeRequest'
      responses:
        '200':
          description: Normalized URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormalizeResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/open:
    post:
      tags:
        - pulsar
      summary: Open a URL immediately
      description: |
        Opens the URL immediately, bypassing the local cache and fetching
        fresh content from the internet.
      operationId: open
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenRequest'
      responses:
        '200':
          description: Page opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/load:
    post:
      tags:
        - pulsar
      summary: Load a URL from storage or internet
      description: |
        Loads a page from local storage if it exists and meets the specified criteria.
        Otherwise, fetches the page from the Internet. Supports expiration, page size
        requirements, and other load options.
      operationId: load
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadRequest'
      responses:
        '200':
          description: Page loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/submit:
    post:
      tags:
        - pulsar
      summary: Submit a URL to the crawl pool
      description: |
        Submits a URL to the URL pool for asynchronous processing. This is a
        non-blocking operation that returns immediately. The URL will be
        processed later in the crawl loop.
      operationId: submit
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitRequest'
      responses:
        '200':
          description: URL submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session identifier
...
  schemas:
    NewSessionRequest:
      type: object
      properties:
        capabilities:
          type: object
          description: Desired capabilities for the session
          additionalProperties: true
    NewSessionResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            sessionId:
              type: string
            capabilities:
              type: object
              additionalProperties: true
    SessionDetails:
      type: object
      properties:
        value:
          type: object
          properties:
            sessionId:
              type: string
            url:
              type: string
            status:
              type: string
              enum: [active, paused, stopped]
            capabilities:
              type: object
              additionalProperties: true
    WebDriverResponse:
      type: object
      properties:
        value:
          description: Response value (can be null for success)
    ErrorResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            error:
              type: string
            message:
              type: string
            stacktrace:
              type: string
    SetUrlRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: URL to navigate to
    UrlResponse:
      type: object
      properties:
        value:
          type: string
          description: Current URL
...
    ScriptRequest:
      type: object
      required:
        - script
      properties:
        script:
          type: string
          description: JavaScript code to execute
        args:
          type: array
          items: {}
          description: Arguments to pass to the script
    ScriptResponse:
      type: object
      properties:
        value:
          description: Script return value
...
    NormalizeRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: The URL to normalize
        args:
          type: string
          description: Optional load arguments (e.g., "-expire 1d")
        toItemOption:
          type: boolean
          default: false
          description: Whether to convert to item load options
    NormalizeResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            spec:
              type: string
              description: The normalized URL specification
            url:
              type: string
              description: The normalized URL
            args:
              type: string
              description: Parsed arguments
            isNil:
              type: boolean
              description: Whether the URL is nil/invalid
    OpenRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: The URL to open immediately
    OpenResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/WebPageResult'
    LoadRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: The URL to load
        args:
          type: string
          description: |
            Optional load arguments. Examples:
            - "-expire 1d" for 1 day expiration
            - "-refresh" for forced refresh
            - "-requireSize 100000" for minimum page size
    LoadResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/WebPageResult'
    SubmitRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: The URL to submit to the crawl pool
        args:
          type: string
          description: Optional load arguments for when the URL is processed
    SubmitResponse:
      type: object
      properties:
        value:
          type: boolean
          description: Whether the URL was submitted successfully
    WebPageResult:
      type: object
      properties:
        url:
          type: string
          description: The page URL
        location:
          type: string
          description: The final location after redirects
        contentType:
          type: string
          description: Content type of the page
        contentLength:
          type: integer
          description: Content length in bytes
        protocolStatus:
          type: string
          description: HTTP protocol status
        isNil:
          type: boolean
          description: Whether the page is nil/invalid